requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
comtypes==1.1.7
decorator==4.4.0
discord.py==1.2.5
future==0.17.1
google-api-core==1.15.0
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-translate==2.0.0
googleapis-common-protos==1.6.0
googletrans==2.4.0
grpcio==1.26.0
idna==2.8
Markdown==2.6.11
multidict==4.7.2
networkx==2.3
numpy==1.16.2
pandas==0.24.2
protobuf==3.11.2
psutil==5.6.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyglet==1.3.2
pypiwin32==223
PySide2==5.13.0
python-dateutil==2.8.0
pytz==2018.9
pywin32==224
requests==2.22.0
rsa==4.0
scipy==1.2.1
selenium==3.141.0
shiboken2==5.13.0
six==1.12.0
soupsieve==1.9.5
urllib3==1.25.7
virtualenv==16.6.0
websockets==6.0
yandex-translater==6.0
yarl==1.4.2